Fixed includes after moving SFTBDDPropertyFormulaAdapter
