+++ src/storm-dft/modelchecker/DFTModularizer.cpp
@@ -1,10 +1,10 @@
 #include <sstream>
 
+#include "storm-dft/adapters/SFTBDDPropertyFormulaAdapter.h"
 #include "storm-dft/builder/DFTBuilder.h"
 #include "storm-dft/modelchecker/DFTModelChecker.h"
 #include "storm-dft/modelchecker/DFTModularizer.h"
 #include "storm-dft/modelchecker/SFTBDDChecker.h"
-#include "storm-dft/modelchecker/SFTBDDPropertyFormulaAdapter.h"
 #include "storm-parsers/api/properties.h"
 #include "storm/api/properties.h"
